requirements.txt
Flask>=3.0.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
deepgaze_pytorch>=1.2.1
einops>=0.7.0
git+https://github.com/openai/CLIP.git
